layout/generic/test/test_bug389630.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=389630
-->
<head>
  <title>Test for Bug 389630</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <style>.fl:first-letter { }
  </style>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=389630">Mozilla Bug 389630</a>
<div id="content" style="direction: rtl;"><table><tr><td width="1"><div id="div" class="fl"><font><span>x y</span></font></div></td></tr></table></div>
<pre id="test">
<script class="testbody" type="text/javascript">
 function boom()
{
/** Test for Bug 389630 **/
  document.getElementById("div").className = "";
  ok(true, "Should not crash");
  SimpleTest.finish();
}
  SimpleTest.requestFlakyTimeout("untriaged");
  setTimeout(boom, 500);
  SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
